Porting some of the PCIe linux tests to UEFI
- Ported PCIe tests 64,66,68 to UEFI.
- Implemented few of the PCIe and SMMU related PAL APIs in VAL.
- Removed unused pal implementations.
- Test numbers modified to be in sequence.
