+++ platform/pal_baremetal/BsaPalBaremetalLib.inf
@@ -35,7 +35,6 @@
   src/pal_timer_wd.c
   src/pal_exerciser.c
   src/pal_iovirt.c
-  src/pal_smmu.c
   src/pal_peripherals.c
   src/pal_misc.c
   src/pal_dma.c
