+++ test_pool/pcie/operating_system/test_os_p041.c
@@ -1,5 +1,5 @@
 /** @file
- * Copyright (c) 2016-2018,2021,2023 Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2016-2018, 2021, 2023, Arm Limited or its affiliates. All rights reserved.
  * SPDX-License-Identifier : Apache-2.0
 
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -19,7 +19,7 @@
 
 #include "val/include/bsa_acs_pcie.h"
 
-#define TEST_NUM   (ACS_PCIE_TEST_NUM_BASE + 66)
+#define TEST_NUM   (ACS_PCIE_TEST_NUM_BASE + 41)
 #define TEST_RULE  "PCI_MM_04"
 #define TEST_DESC  "NP type-1 PCIe supp 32-bit only       "
 
@@ -93,7 +93,7 @@ payload(void)
 }
 
 uint32_t
-os_p066_entry(uint32_t num_pe)
+os_p041_entry(uint32_t num_pe)
 {
     uint32_t status = ACS_STATUS_FAIL;
