+++ uefi_app/BsaAcs.inf
@@ -1,5 +1,5 @@
 ## @file
-#  Copyright (c) 2016-2023 Arm Limited or its affiliates. All rights reserved.
+#  Copyright (c) 2016-2023, Arm Limited or its affiliates. All rights reserved.
 #  SPDX-License-Identifier : Apache-2.0
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
@@ -104,6 +104,9 @@
   ../test_pool/pcie/operating_system/test_os_p037.c
   ../test_pool/pcie/operating_system/test_os_p038.c
   ../test_pool/pcie/operating_system/test_os_p039.c
+  ../test_pool/pcie/operating_system/test_os_p040.c
+  ../test_pool/pcie/operating_system/test_os_p041.c
+  ../test_pool/pcie/operating_system/test_os_p042.c
   ../test_pool/smmu/operating_system/test_os_i001.c
   ../test_pool/smmu/operating_system/test_os_i002.c
   ../test_pool/smmu/operating_system/test_os_i003.c
