+++ val/include/bsa_acs_pcie.h
@@ -1,5 +1,5 @@
 /** @file
- * Copyright (c) 2016-2020,2021,2022 Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2016-2023, Arm Limited or its affiliates. All rights reserved.
  * SPDX-License-Identifier : Apache-2.0
 
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -258,14 +258,14 @@ uint32_t os_p036_entry(uint32_t num_pe);
 uint32_t os_p037_entry(uint32_t num_pe);
 uint32_t os_p038_entry(uint32_t num_pe);
 uint32_t os_p039_entry(uint32_t num_pe);
+uint32_t os_p040_entry(uint32_t num_pe);
+uint32_t os_p041_entry(uint32_t num_pe);
+uint32_t os_p042_entry(uint32_t num_pe);
 
 /* Linux test */
 uint32_t os_p061_entry(uint32_t num_pe);
 uint32_t os_p062_entry(uint32_t num_pe);
+uint32_t os_p063_entry(uint32_t num_pe);
 uint32_t os_p064_entry(uint32_t num_pe);
-uint32_t os_p065_entry(uint32_t num_pe);
-uint32_t os_p066_entry(uint32_t num_pe);
-uint32_t os_p067_entry(uint32_t num_pe);
-uint32_t os_p068_entry(uint32_t num_pe);
 
 #endif
